Security: Harden WebSocket, DB queries, and Expo cleartext traffic
- Explicitly disable cleartext traffic in frontend app.json
- Update WebSocket message handling to use Pydantic TypeAdapter and catch specific exceptions (ValidationError, PyJWTError)
- Replace generic try/excepts with precise exceptions in websocket.py
- Ensure parameter safety in chat_repo and memory_repo execute_query calls by correctly casting variables.

+++ backend/app/infrastructure/repositories/memory_repo.py
@@ -50,8 +50,8 @@ class MemoryRepository:
         query = """
             SELECT * FROM match_memories(
                 query_embedding := $1::vector,
-                match_threshold := $2,
-                match_count := $3,
+                match_threshold := $2::float,
+                match_count := $3::int,
                 p_user_id := $4::uuid,
                 p_agent_id := $5::uuid,
                 p_room_id := $6::uuid
